pyproject.toml
[tool.poetry]
name = "aqt-connector"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "aqt_connector" },
    { include = "tests", format = "sdist" },
]


[tool.poetry.dependencies]
python = ">=3.9,<3.14"
httpx = "^0.27.2"
typer = "^0.13.0"
auth0-python = "^4.7.2"
qrcode = "^8.0"
tomli = "^2.1.0"
pydantic = "^2.10.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
mypy = "^1.13.0"
ruff = "^0.9.9"
typos = "^1.30.1"


[tool.poetry.group.test.dependencies]
pytest-playwright = "^0.7.0"
interrogate = "^1.7.0"



[tool.poetry.group.docs.dependencies]
pdoc3 = "^0.11.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
# Allow lines to be as long as 120.
line-length = 120

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
